bot/requirements.txt
aiogram==3.13.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
python-dotenv==1.0.1
aiohttp-socks==0.10.1
